+++ api-colegio/src/main/java/com/apolo/modulos/roles/model/Rol.java
@@ -12,7 +12,7 @@ public class Rol {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @Size(min=2, message = "minimo 2")
     @NotNull
@@ -26,17 +26,17 @@ public class Rol {
 
     }
 
-    public Rol(Integer id, String nombre, String credencial) {
+    public Rol(Long id, String nombre, String credencial) {
         this.id = id;
         this.nombre = nombre;
         this.credencial = credencial;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
