+++ api-colegio/src/main/java/com/apolo/modulos/usuarios/model/TokenActivacionUsuario.java
@@ -1,7 +1,5 @@
 package com.apolo.modulos.usuarios.model;
 
-import com.apolo.modulos.usuarios.model.Usuario;
-
 import javax.persistence.*;
 
 
@@ -17,7 +15,7 @@ public class TokenActivacionUsuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String token;
 
@@ -42,11 +40,11 @@ public class TokenActivacionUsuario {
         this.usuario = usuario;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
